story update: sidebar entry
